+++ tests/Response/DTO/Item/UserTest.php
@@ -1,27 +1,24 @@
 <?php
 
-namespace Response\DTO;
+namespace Response\DTO\Item;
 
-use AstrobinWs\Response\DTO\Item\Image;
+use AstrobinWs\Response\DTO\Item\User;
 use DG\BypassFinals;
 use PHPUnit\Framework\TestCase;
 
-class ImageTest extends TestCase
+class UserTest extends TestCase
 {
     public static array $expectedProperties = [
-        'description',
-        'subjects',
-        'title',
-        'uploaded',
-        'url_gallery',
-        'url_hd',
-        'url_histogram',
-        'url_regular',
-        'url_skyplot',
-        'url_thumb',
-        'user'
+        'about',
+        'avatar',
+        'hobbies',
+        'id',
+        'image_count',
+        'job',
+        'language',
+        'username',
+        'website'
     ];
-
     protected function setUp(): void
     {
         BypassFinals::enable();
@@ -29,7 +26,7 @@ class ImageTest extends TestCase
 
     public function testPropertiesExist(): void
     {
-        $reflexion = new \ReflectionClass((new Image()));
+        $reflexion = new \ReflectionClass((new User()));
         $props = array_map(static fn(\ReflectionProperty $prop) => $prop->getName(), $reflexion->getProperties());
         sort($props);
         $this->assertEquals(self::$expectedProperties, $props);
